refactor CLI class into smaller class
